fix: pull shared logic into concore_base, fix docker path/cleanup bugs
